// src/services/api/gen/types.gen.ts
export type BusinessProfile = {
    name: string;
    description?: string | null;
    logo?: string | null;
    readonly logo_url: string | null;
    contact_no: string;
    whatsapp_no?: string | null;
    tagline?: string | null;
    opening_time: string;
    closing_time: string;
    establishment: string;
    payment_mode: string;
    email: string;
    website?: string | null;
    plan: string;
};